Move unit tests to tests/dowel (#4)
This PR moves unit tests modules to `tests/dowel`. This makes it so
that the unit tests paths for a module are easy to predict. For
instance, the tests for the module `dowel.csv_output` will now live at
`tests.dowel.test_csv_output` rather than `tests.test_csv_output`.